% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_dataset}
\alias{example_dataset}
\title{Example dataset}
\format{
\subsection{\code{example_dataset}}{

A tibblwe with 33 columns:
\describe{
\item{variable}(IL6)
\item{description}(mesurements of IL6 levels)
}
}
}
\source{
data generated with the simstudy package
}
\usage{
data(example_dataset)
}
\description{
Example dataset
}
\keyword{datasets}
