% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_scatterplot.R
\name{make_scatterplot}
\alias{make_scatterplot}
\title{create a random scatterplot}
\usage{
make_scatterplot(tsvfuke)
}
\arguments{
\item{tsvfuke}{string indicating path to a tsv dataset}
}
\value{
a plot object with a scatterplot from randomly selected variables
}
\description{
create a random scatterplot
}
