% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_dataset.R
\name{summarise_dataset}
\alias{summarise_dataset}
\title{Summarises a DataSet}
\usage{
summarise_dataset(tsv_file)
}
\arguments{
\item{tsv_fileA}{A string with the path to a TSV file with data we would like to import}
}
\value{
A tibble with 4 variables, summarising average, min and max of each variable
}
\description{
Summarises a DataSet
}
\examples{
\dontrun {
summarise_dataset("/my/path/to/dataset.tsv")
}
}
