% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_dataset.R
\name{summarize_dataset}
\alias{summarize_dataset}
\title{Summarises a Dataset}
\usage{
summarize_dataset(tsvfuke)
}
\arguments{
\item{tsvfuke}{A string with the path to a TSV file with data we would like to import}
}
\value{
A tibble with 4 variables, summarising average, min and max of each variable
}
\description{
Summarises a Dataset
}
\examples{
\dontrun{
summarize_dataset("/my/path/to/dataset.tsv")
}
}
