% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_scatterplot.R
\name{make_scatterplot}
\alias{make_scatterplot}
\title{Create a random scatterplot}
\usage{
make_scatterplot(tsv_file)
}
\arguments{
\item{tsv_file}{String indicating path to a tsv file}
}
\value{
A plot object with a scatterplot
}
\description{
Create a random scatterplot
}
