% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_scatterplot.R
\name{make_scatterplot}
\alias{make_scatterplot}
\title{Creates a random scatterplot}
\usage{
make_scatterplot(tsv_file)
}
\arguments{
\item{tsv_file}{A string indicating path to a TSV dataset}
}
\value{
A plot object with a scatterplot from randomly chosen variables
}
\description{
Creates a random scatterplot
}
\examples{
\dontrun {
make_scatterplot("/my/path/to/data.tsv")
}
}
