% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_dataset.R
\name{summarize_dataset}
\alias{summarize_dataset}
\title{Summarises a Dataset}
\usage{
summarize_dataset(tsv_file)
}
\arguments{
\item{tsv_file}{A TSV file with data we would like to import}
}
\value{
A Tibble with 4 variables, summarizing average, min and max of each variable
}
\description{
Summarises a Dataset
}
